Crate re_global_context

State that is shared with most top-level crates.

Modules

app_options 🔒

blueprint_id 🔒

command_sender 🔒

contents 🔒

file_dialog 🔒

item 🔒

recording_or_table 🔒

Structs

AppOptions

Global options for the viewer.

BlueprintId

A unique id for a type of Blueprint contents.

CommandReceiver

Receiver for the CommandSender

CommandSender

Sender that queues up the execution of commands.

GlobalContext

Application context that is shared across all parts of the viewer.

Enums

Contents

ContentsName

The name of a Contents.

DisplayMode

Which display mode are we currently in?

Item

One “thing” in the UI.

RecordingOrTable

SystemCommand

Commands used by internal system components

Traits

BlueprintIdRegistry

SystemCommandSender

Interface for sending SystemCommand messages.

Functions

blueprint_id_to_tile_id

command_channel

Creates a new command channel.

resolve_mono_instance_path

If the given path refers to the first element of an instance with a single element, resolve to a unindexed entity path.

resolve_mono_instance_path_item

If the given item refers to the first element of an instance with a single element, resolve to a unindexed entity path.

santitize_file_name

Replace “dangerous” characters by a safe one.

Type Aliases

ContainerId

ViewId